Add logic for creating one small div grid box
This commit will include the creation of one small
gridbox, which will be appended to it's line box
with display:flex which will be appended to the
main div container with display:inline-block.

The next step is to create loops for both the
small boxes to create 16 in one line, as well as
a loop for the line box to create and append 16
lines to the main div container, and prevent
manually writing the code for each.