Add correct logic for user input selection
Linked functions to remove old grid and create
a new one in the container, based on user input.

The do while loop ensures that:
- if a user presses OK without entering any text
or presses Cancel, this will provide empty string
or null respetively.
- The loop will exit and stop asking the user,
avoiding the infinity loop i previously created.
- Using Number function on the input, if user
types string, the number function provides NaN.
Otherwise, will provide 0 or a number.
- Included a conditional that checks if the input
is within the range and is not NaN!